Fix (work around) segmentation fault in logger for OpenMP
When spawning multiple threads (OpenMP) for hicann-system, the old usage
of the logger yields a segmentation fault at the end of the execution.
This is triggered by multiple destructions of the last logger's Message
object. This commit works around by avoiding the last call to ~Message.
Caveat: It's ugly and the last logger message will be lost.

Change-Id: I3ba890e78791c64840b8a00369e5ce3a76e48543